requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
praw==7.1.0
prawcore==1.4.0
PyJWT==1.7.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
twilio==6.44.0
update-checker==0.17
websocket-client==0.57.0
